Some module restructuring.
