--- Seagull-APIs/Seagull-Public-APIs/pom.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    
-    PROJECT
-        Name
-            Seagull-Service-APIs
-        
-        Description
-            Seagull - Intended to be a very simplistic service platform.
-            
-            The idea here is to define a service platform that says
-            nothing about how services communicate with each other.
-            This defines APIs and not to many of those, that can be
-            implemented with whatever protocol. What protocol is used
-            depends on what implementation you make available on
-            the classpath. Implementations are fetched using
-            SGProviderLookup (currently ServiceLoader is used).
-        
-    COPYRIGHTS
-        Copyright (C) 2023 by Tommy Bengt Svensson All rights reserved.
-        
-    LICENSE
-        Apache 2.0 (Open Source)
-        
-        Licensed under the Apache License, Version 2.0 (the "License");
-        you may not use this file except in compliance with the License.
-        You may obtain a copy of the License at
-        
-          http://www.apache.org/licenses/LICENSE-2.0
-        
-        Unless required by applicable law or agreed to in writing, software
-        distributed under the License is distributed on an "AS IS" BASIS,
-        WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-        See the License for the specific language governing permissions and
-        limitations under the License.
-        
-    AUTHORS
-        tommy ()
-            Changes:
-            2023-11-04: Created!
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>se.natusoft.seagull</groupId>
-        <artifactId>Seagull-APIs</artifactId>
-        <version>1.0.0-BC11</version>
-    </parent>
-
-    <artifactId>Seagull-Public-APIs</artifactId>
-    <version>1.0.0-BC11</version>
-
-    <dependencies>
-
-    </dependencies>
-
-</project>
