+++ Seagull-Platform/src/main/groovy/se/natusoft/seagull/api/SGService.groovy
