+++ Seagull-Platform/src/main/groovy/se/natusoft/seagull/api/model/SGCRUD.groovy
