+++ Seagull-Platform/src/test/groovy/se/natusoft/seagull/exceptions/TestExceptions.groovy
